Remove FinalCommand
